# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Alsaqi Internal Audit API
  version: 1.0.0
  description: >-
    OpenAPI contract for the Alsaqi backend. Operations are derived from the
    routes registered under the unified API base in src/routes/v1/index.ts.
    Paths are expressed on the unversioned /api base per the unified frontend
    contract; the versioned /api/v1 form resolves to the same handlers. The
    Docs_Endpoint (GET /docs) is intentionally excluded from this contract.
servers:
  - url: /api
    description: Unversioned API base path (unified contract)
  - url: /api/v1
    description: Versioned API base path (backward compatible)
paths:
...
  '/notifications':
    get:
      summary: GET /notifications
      description: >-
        Paginated notifications for the authenticated user. The notification
        items are carried inside data and pagination metadata in meta.pagination.
      operationId: get_notifications
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Success_Envelope with data (notification collection) and meta.pagination.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success_Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
        '401':
          description: Unauthenticated request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error_Envelope'
  '/notifications/mark-all-read':
    put:
      summary: PUT /notifications/mark-all-read
      description: >-
        Marks every unread notification of the authenticated user as read.
        Returns the count of notifications changed to read.
      operationId: put_notifications_mark_all_read
      responses:
        '200':
          description: Success_Envelope whose data.updated is the number of notifications marked read.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success_Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          updated:
                            type: integer
        '401':
          description: Unauthenticated request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error_Envelope'
  '/notifications/mark-read':
    put:
      summary: PUT /notifications/mark-read
      description: >-
        Bulk-marks the listed notifications of the authenticated user as read.
        Returns the count of notifications actually changed to read.
      operationId: put_notifications_mark_read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [notification_ids]
              properties:
                notification_ids:
                  type: array
                  items:
                    oneOf:
                      - type: string
                      - type: integer
      responses:
        '200':
          description: Success_Envelope whose data.updated is the number of notifications marked read.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success_Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          updated:
                            type: integer
        '400':
          description: notification_ids missing or not an array.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error_Envelope'
        '401':
          description: Unauthenticated request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error_Envelope'
  '/notifications/unread-count':
    get:
      summary: GET /notifications/unread-count
      description: Count of unread notifications for the authenticated user.
      operationId: get_notifications_unread_count
      responses:
        '200':
          description: Success_Envelope whose data is the unread count value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success_Envelope'
        '401':
          description: Unauthenticated request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error_Envelope'
  '/notifications/{id}':
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      summary: DELETE /notifications/{id}
      description: >-
        Dismisses (soft-deletes) a notification for the authenticated user; the
        row is not physically deleted from the database.
      operationId: delete_notifications_id
      responses:
        '200':
          description: "Success_Envelope with data { success: true }."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success_Envelope'
        '401':
          description: Unauthenticated request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error_Envelope'
  '/notifications/{id}/read':
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      summary: PUT /notifications/{id}/read
      description: Marks a single notification of the authenticated user as read.
      operationId: put_notifications_id_read
      responses:
        '200':
          description: "Success_Envelope with data { success: true }."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success_Envelope'
        '401':
          description: Unauthenticated request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error_Envelope'
...
components:
  schemas:
    Success_Envelope:
      type: object
      description: >-
        Standard success envelope produced by createSuccessResponse. Successful
        responses always carry success=true, a data payload, and a meta object.
        Paginated collections place their pagination metadata in meta.pagination.
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          description: The response payload. Shape varies per endpoint.
          nullable: true
        meta:
          $ref: '#/components/schemas/Meta'
    Error_Envelope:
      type: object
      description: >-
        Standard error envelope produced by createErrorResponse. Error responses
        (HTTP status >= 400) always carry success=false, data=null, and an error
        object with a non-empty code and message.
      required:
        - success
        - data
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        data:
          type: object
          nullable: true
          enum: [null]
        error:
          $ref: '#/components/schemas/Error_Detail'
        meta:
          $ref: '#/components/schemas/Meta'
    Error_Detail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code (non-empty).
        message:
          type: string
          description: Human-readable error message (non-empty).
        details:
          description: Optional additional error context (e.g. validation issues).
          nullable: true
    Meta:
      type: object
      description: Response metadata. Includes pagination for paginated collections.
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    Notification:
      type: object
      description: A notification feed item for the current user.
      required:
        - event_type
        - description
        - related_module
        - date
      properties:
        id:
          oneOf:
            - type: string
            - type: integer
        recipient_row_id:
          type: string
        user_id:
          type: integer
        event_type:
          type: string
        title:
          type: string
          nullable: true
        description:
          type: string
        related_module:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [Read, Unread]
        is_read:
          type: boolean
        read_at:
          type: string
          nullable: true
        link:
          type: string
        actor_id:
          type: string
        entity_id:
          type: string
        entity_type:
          type: string
        data:
          type: object
          additionalProperties: true